Cleaned up the anyval type conversion stuff
